# Typesafe map intrinsics