--- a/stdlib/mexpr/ast.mc
+++ b/stdlib/mexpr/ast.mc
@@ -754,6 +754,7 @@ lang MapAst = ConstAst
   | CMapFindApplyOrElse {}
   | CMapBindings {}
   | CMapSize {}
+  | CMapIsEmpty {}
   | CMapMem {}
   | CMapAny {}
   | CMapMap {}