This PR re-implements the map intrinsics in a typesafe way (without `Obj`). It was done with help from @elegios.

Some preliminary benchmarks (run on my laptop while I didn't touch it).
*Before change:*
```
Bootstrapping the Miking compiler (1st round, might take a few minutes)

real	1m4.069s
user	1m2.929s
sys	0m0.887s
Bootstrapping the Miking compiler (2nd round, might take some more time)

real	0m12.359s
user	0m11.520s
sys	0m0.620s

$ for i in {1..20}; do time mi compile src/main/mi.mc; done                       lingmar@lingmar-mbp
mi compile src/main/mi.mc  11.54s user 0.61s system 99% cpu 12.248 total
mi compile src/main/mi.mc  11.53s user 0.63s system 99% cpu 12.171 total
mi compile src/main/mi.mc  11.51s user 0.61s system 99% cpu 12.149 total
mi compile src/main/mi.mc  11.51s user 0.61s system 99% cpu 12.156 total
mi compile src/main/mi.mc  11.51s user 0.62s system 99% cpu 12.155 total
mi compile src/main/mi.mc  11.60s user 0.62s system 99% cpu 12.255 total
mi compile src/main/mi.mc  11.53s user 0.62s system 99% cpu 12.218 total
mi compile src/main/mi.mc  11.55s user 0.62s system 99% cpu 12.198 total
mi compile src/main/mi.mc  11.57s user 0.62s system 99% cpu 12.210 total
mi compile src/main/mi.mc  11.52s user 0.62s system 99% cpu 12.173 total
mi compile src/main/mi.mc  11.54s user 0.63s system 99% cpu 12.193 total
mi compile src/main/mi.mc  11.55s user 0.62s system 99% cpu 12.222 total
mi compile src/main/mi.mc  11.49s user 0.62s system 99% cpu 12.150 total
mi compile src/main/mi.mc  11.55s user 0.62s system 99% cpu 12.196 total
mi compile src/main/mi.mc  11.51s user 0.62s system 99% cpu 12.156 total
mi compile src/main/mi.mc  11.61s user 0.62s system 99% cpu 12.247 total
mi compile src/main/mi.mc  11.51s user 0.62s system 99% cpu 12.154 total
mi compile src/main/mi.mc  11.52s user 0.62s system 99% cpu 12.166 total
mi compile src/main/mi.mc  11.57s user 0.63s system 99% cpu 12.235 total
mi compile src/main/mi.mc  11.51s user 0.61s system 99% cpu 12.145 total
```

*After change*
```
Bootstrapping the Miking compiler (1st round, might take a few minutes)


real	1m28.254s
user	1m26.571s
sys	0m1.105s
Bootstrapping the Miking compiler (2nd round, might take some more time)

real	0m12.991s
user	0m12.001s
sys	0m0.651s

for i in {1..20}; do time mi compile src/main/mi.mc; done                       lingmar@lingmar-mbp
mi compile src/main/mi.mc  11.90s user 0.63s system 99% cpu 12.643 total
mi compile src/main/mi.mc  11.90s user 0.64s system 99% cpu 12.583 total
mi compile src/main/mi.mc  11.89s user 0.64s system 99% cpu 12.578 total
mi compile src/main/mi.mc  11.90s user 0.64s system 99% cpu 12.572 total
mi compile src/main/mi.mc  11.93s user 0.64s system 99% cpu 12.608 total
mi compile src/main/mi.mc  11.89s user 0.64s system 99% cpu 12.574 total
mi compile src/main/mi.mc  11.85s user 0.63s system 99% cpu 12.514 total
mi compile src/main/mi.mc  11.78s user 0.63s system 99% cpu 12.450 total
mi compile src/main/mi.mc  11.87s user 0.63s system 99% cpu 12.529 total
mi compile src/main/mi.mc  11.86s user 0.64s system 99% cpu 12.522 total
mi compile src/main/mi.mc  11.89s user 0.65s system 99% cpu 12.572 total
mi compile src/main/mi.mc  11.86s user 0.64s system 99% cpu 12.530 total
mi compile src/main/mi.mc  12.13s user 0.68s system 98% cpu 12.947 total
mi compile src/main/mi.mc  11.85s user 0.64s system 99% cpu 12.531 total
mi compile src/main/mi.mc  11.88s user 0.65s system 99% cpu 12.547 total
mi compile src/main/mi.mc  11.82s user 0.64s system 99% cpu 12.489 total
mi compile src/main/mi.mc  11.87s user 0.64s system 99% cpu 12.534 total
mi compile src/main/mi.mc  11.85s user 0.64s system 99% cpu 12.510 total
mi compile src/main/mi.mc  11.92s user 0.64s system 99% cpu 12.584 total
mi compile src/main/mi.mc  11.88s user 0.63s system 99% cpu 12.540 total
```

So it seems the changes make things a bit slower, we might want to investigate why before merging.