@elegios 2021-09-13
Did you try the thing with having `Mmap.Map` contain a nested module of the correct signature instead of using `include`?