@lingmar 2021-09-13
Do you mean something along the line of:
```ocaml
  module type Map = sig
    type k

    type v

    module type MapInner = sig
      include Map.S with type key := k
    end

    val theMap : v MapInner.t

    val cmp : k -> k -> int
  end
```
this doesn't type check but with something like this the inner module wouldn't need to be copied possibly.