@lingmar 2021-09-13
Ah, no, I guess you meant:
```
  module type Map = sig
    type k

    type v

    module MapInner : Map.S with type key := k

    val theMap : v MapInner.t

    val cmp : k -> k -> int
  end
```