@lingmar 2021-09-13
I implemented the latter but saw no difference in runtime from the first version (using `include`).