@elegios 2021-09-14
Interesting. I guess the difference must then be either because of the extra work in `eq` and `cmp` (I think the lazy sequences of ocaml aren't entirely free) or the extra indirection (I assume there is one at least, one to the module we pass around, then the reference to the actual value).