@lingmar 2021-09-14
Closing since the runtime overheads of the changes are too large. We should look into another map implementation in the future.